chore: restructured code
