+++ src/interfaces/data.ts
@@ -1,0 +1,9 @@
+export interface Joke {
+	categories: string[] | never[];
+	created_at: string;
+	icon_url: string;
+	id: string;
+	updated_at: string;
+	url: string;
+	value: string;
+}
