+++ src/interfaces/props.ts
@@ -1,0 +1,6 @@
+export interface CreateRangeProps {
+	from?: number;
+	to: number;
+	step?: number;
+	length?: number;
+}
